payrollhq/organizations/views.py
from rest_framework import viewsets
from .models import Organization

class OrganizationViewSet(viewsets.ModelViewSet):
    """
    ViewSet for managing organizations
    """
    queryset = Organization.objects.all()
    
    def get_queryset(self):
        # In a real implementation, filter by user's organization
        return super().get_queryset()