payrollhq/requirements.txt
# PayrollHQ Requirements
# Core Django and DRF
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0.0

# Database
psycopg2-binary>=2.9.0

# Configuration
python-decouple>=3.8

# Date and Time
python-dateutil>=2.8.0

# Decimal arithmetic
decimal>=1.2.0

# Email
django-anymail>=10.0

# Testing
pytest>=7.0.0
pytest-django>=4.5.0
factory-boy>=3.2.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Production
gunicorn>=20.1.0
whitenoise>=6.4.0

# Monitoring
sentry-sdk>=1.25.0

# Excel export (for compliance reports)
openpyxl>=3.1.0

# PDF generation (for payslips)
reportlab>=4.0.0